requirements.txt
absl-py==2.1.0
africastalking==1.2.7
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
bleach==6.1.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
contextlib2==21.6.0
contourpy==1.2.0
crispy-bootstrap5==2024.2
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
Django
django-crispy-forms==2.1
dm-tree==0.1.8
executing==2.0.1
flatbuffers==24.3.7
fonttools==4.50.0
gast==0.5.4
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.1
google-generativeai==0.4.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.62.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
jedi==0.19.1
kiwisolver==1.4.5
libclang==16.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.7
numpy==1.26.4
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==24.0
parso==0.8.3
pillow==10.2.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.17.2
pymemcache==4.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-memcached==1.62
python3-memcached==1.51
requests==2.31.0
rich==13.7.1
rsa==4.9
schema==0.7.5
setuptools==69.2.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
stack-data==0.6.3
termcolor==2.4.0
tinycss2==1.2.1
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.1
wheel==0.43.0
wrapt==1.16.0
whitenoise==6.6.0
daphne
channels
redis
channels_redis
